Normalize movement arguments to [-1.0,1.0] range
Every PTZ protocol has a different encoding, some using integers, some
using floating point, and different ranges for each. This poses
difficulties for a generic controller that needs to manage all of them.
For the generic code, normalize all the movement ranges to be in the
range [-1.0,1.0] or [0.0,1.0] floating point, and make it the
responsibility of the backend driver to translate those ranges to the
range of the actual camera.

For now the VISCA protocol uses scaling values hardcoded to the Sony
SRG-120DH. A future change will fix this to allow the
pan/tilt/focus/zoom range to be set per-camera model.

+++ src/ptz-onvif.cpp
@@ -668,14 +668,14 @@ void PTZOnvif::pantilt(double pan, double tilt)
 	}
 }
 
-void PTZOnvif::pantilt_abs(int pan, int tilt)
+void PTZOnvif::pantilt_abs(double pan, double tilt)
 {
 	OnvifPTZService c;
 	c.AbsoluteMove(m_PTZAddress, username, password, m_selectedMedia.token,
 		       pan, tilt, 0.0);
 }
 
-void PTZOnvif::pantilt_rel(int pan, int tilt)
+void PTZOnvif::pantilt_rel(double pan, double tilt)
 {
 	OnvifPTZService c;
 	c.RelativeMove(m_PTZAddress, username, password, m_selectedMedia.token,
@@ -701,7 +701,7 @@ void PTZOnvif::zoom(double speed)
 	}
 }
 
-void PTZOnvif::zoom_abs(int pos)
+void PTZOnvif::zoom_abs(double pos)
 {
 	OnvifPTZService c;
 	c.AbsoluteMove(m_PTZAddress, username, password, m_selectedMedia.token,
